+++ src/dapp/interface.ts
@@ -4,12 +4,13 @@ import {SignableMessage} from "../signableMessage";
 export interface IDappProvider {
     init(): Promise<boolean>;
     login(options?: {callbackUrl?: string; token?: string; addressIndex?: number}): Promise<string>;
-    logout(): Promise<boolean>;
+    logout(options?: {callbackUrl?: string}): Promise<boolean>;
     getAddress(): Promise<string>;
     isInitialized(): boolean;
     isConnected(): Promise<boolean>;
     sendTransaction(transaction: Transaction, options?: {callbackUrl?: string}): Promise<Transaction>;
     signTransaction(transaction: Transaction, options?: {callbackUrl?: string}): Promise<Transaction>;
+    signTransactions(transaction: Array<Transaction>, options?: {callbackUrl?: string}): Promise<Array<Transaction>>;
     signMessage(transaction: SignableMessage, options?: {callbackUrl?: string}): Promise<SignableMessage>;
 }