+++ src/dapp/walletProvider.ts
@@ -1,55 +1,51 @@
-import { IDappProvider, IDappMessageEvent } from "./interface";
+import qs from "qs";
+import { IDappProvider } from "./interface";
 import {
-    DAPP_MESSAGE_INIT,
-    DAPP_DEFAULT_TIMEOUT,
-    DAPP_MESSAGE_IS_CONNECTED,
-    DAPP_MESSAGE_GET_ADDRESS,
-    DAPP_MESSAGE_CONNECT_URL,
-    DAPP_MESSAGE_SEND_TRANSACTION_URL,
-    DAPP_MESSAGE_SIGN_TRANSACTION_URL,
-    DAPP_MESSAGE_LOG_OUT
+    WALLET_PROVIDER_CALLBACK_PARAM,
+    WALLET_PROVIDER_CALLBACK_PARAM_TX_SIGNED,
+    WALLET_PROVIDER_CONNECT_URL,
+    WALLET_PROVIDER_DISCONNECT_URL,
+    WALLET_PROVIDER_SEND_TRANSACTION_URL,
+    WALLET_PROVIDER_SIGN_TRANSACTION_URL,
 } from "./constants";
-import { mainFrameStyle } from "./dom";
 import { Transaction } from "../transaction";
 import { SignableMessage } from "../signableMessage";
-import { ErrNotImplemented } from "../errors";
+import {ErrInvalidTxSignReturnValue, ErrNotImplemented} from "../errors";
+import {Nonce, Balance, Address, GasPrice, GasLimit, TransactionPayload, ChainID, TransactionVersion} from "../";
+import {Signature} from "../signature";
+
+interface TransactionMessage {
+    receiver: string;
+    value: string;
+    gasPrice?: number;
+    gasLimit?: number;
+    data?: string;
+    nonce?: number;
+}
 
 export class WalletProvider implements IDappProvider {
     walletUrl: string;
-    customId: string;
-    mainFrame: HTMLIFrameElement | undefined;
 
     /**
      * Creates a new WalletProvider
      * @param walletURL
-     * @param customId
      */
-    constructor(walletURL: string = '', customId = '') {
+    constructor(walletURL: string = '') {
         this.walletUrl = walletURL;
-        this.customId = customId;
-        this.attachMainFrame();
-        let _ = this.init();
     }
 
     /**
      * Waits for the wallet iframe to ping that it has been initialised
      */
     async init(): Promise<boolean> {
-        if (this.isInitialized()) {
-            return true;
-        }
-        return this.waitForRemote();
+        return true;
     }
 
     /**
      * Returns if the wallet iframe is up and running
      */
     isInitialized(): boolean {
-        if (!this.mainFrame) {
-            return false;
-        }
-
-        return this.mainFrame.dataset.initialized === "true";
+        return true;
     }
 
     /**
@@ -57,184 +53,45 @@ export class WalletProvider implements IDappProvider {
      *  and has the wallet session active
      */
     async isConnected(): Promise<boolean> {
-        if (!this.mainFrame) {
-            return false;
-        }
-
-        const { contentWindow } = this.mainFrame;
-        if (!contentWindow) {
-            return false;
-        }
-
-        return new Promise((resolve, reject) => {
-            console.log("postMessage", DAPP_MESSAGE_IS_CONNECTED);
-
-            contentWindow.postMessage({
-                type: DAPP_MESSAGE_IS_CONNECTED,
-            }, this.walletUrl);
-
-            const timeout = setTimeout(_ => reject('window not responding'), 5000);
-
-            const isConnected = (ev: IDappMessageEvent) => {
-                console.log("event", "isConnected", ev);
-
-                if (!this.isValidWalletSource(ev.origin)) {
-                    return;
-                }
-
-                const { data } = ev;
-                if (data.type !== DAPP_MESSAGE_IS_CONNECTED) {
-                    return;
-                }
-
-                clearTimeout(timeout);
-                window.removeEventListener('message', isConnected.bind(this));
-                return resolve(data.data);
-            };
-
-            window.addEventListener('message', isConnected);
-        });
+        return false;
     }
 
     /**
      * Fetches the login hook url and redirects the client to the wallet login.
      */
     async login(options?: { callbackUrl?: string; token?: string }): Promise<string> {
-        if (!this.mainFrame) {
-            return '';
+        let callbackUrl = `callbackUrl=${window.location.href}`;
+        if (options && options.callbackUrl) {
+            callbackUrl = `callbackUrl=${options.callbackUrl}`;
         }
 
-        const { contentWindow } = this.mainFrame;
-        if (!contentWindow) {
-            console.warn("something went wrong, main wallet iframe does not contain a contentWindow");
-            return '';
+        let token = '';
+        if (options && options.token) {
+            token = `&token=${options.token}`;
         }
 
-        return new Promise<string>((resolve, reject) => {
-            console.log("postMessage", DAPP_MESSAGE_CONNECT_URL);
-
-            contentWindow.postMessage({
-                type: DAPP_MESSAGE_CONNECT_URL,
-            }, this.walletUrl);
-
-            const timeout = setTimeout(_ => reject('connect url not responding'), 5000);
-            const connectUrl = (ev: IDappMessageEvent) => {
-                console.log("event", "connectUrl", ev);
-
-                if (!this.isValidWalletSource(ev.origin)) {
-                    return;
-                }
-
-                const { data } = ev;
-                if (data.type !== DAPP_MESSAGE_CONNECT_URL) {
-                    return;
-                }
-
-                clearTimeout(timeout);
-                window.removeEventListener('message', connectUrl.bind(this));
-                return resolve(data.data.toString());
-            };
-
-            window.addEventListener('message', connectUrl);
-        }).then((connectionUrl: string) => {
-            let callbackUrl = `callbackUrl=${window.location.href}`;
-            if (options && options.callbackUrl) {
-                callbackUrl = `callbackUrl=${options.callbackUrl}`;
-            }
-            let token = '';
-            if (options && options.token) {
-                token = `&token=${options.token}`;
-            }
-            window.location.href = `${this.baseWalletUrl()}${connectionUrl}?${callbackUrl}${token}`;
-            return window.location.href;
-        }).catch(_ => {
-            return '';
-        });
+        window.location.href = `${this.baseWalletUrl()}${WALLET_PROVIDER_CONNECT_URL}?${callbackUrl}${token}`;
+        return window.location.href;
     }
 
     /**
     * Fetches the logout hook url and redirects the client to the wallet logout.
     */
-    async logout(): Promise<boolean> {
-        if (!this.mainFrame) {
-            return false;
+    async logout(options?: { callbackUrl?: string }): Promise<boolean> {
+        let callbackUrl = `callbackUrl=${window.location.href}`;
+        if (options && options.callbackUrl) {
+            callbackUrl = `callbackUrl=${options.callbackUrl}`;
         }
 
-        const { contentWindow } = this.mainFrame;
-        if (!contentWindow) {
-            console.warn("something went wrong, main wallet iframe does not contain a contentWindow");
-            return false;
-        }
-
-        return new Promise<boolean>((resolve, reject) => {
-            console.log("postMessage", DAPP_MESSAGE_LOG_OUT);
-
-            contentWindow.postMessage({
-                type: DAPP_MESSAGE_LOG_OUT,
-            }, this.walletUrl);
-
-            const timeout = setTimeout(_ => reject('logout url not responding'), 5000);
-            const logout = (ev: IDappMessageEvent) => {
-                console.log("event", "logouturl", ev);
-
-                if (!this.isValidWalletSource(ev.origin)) {
-                    return;
-                }
-
-                const { data } = ev;
-                if (data.type !== DAPP_MESSAGE_LOG_OUT) {
-                    return;
-                }
-
-                clearTimeout(timeout);
-                window.removeEventListener('message', logout.bind(this));
-                return resolve(data.data);
-            };
-
-            window.addEventListener('message', logout);
-        });
+        window.location.href = `${this.baseWalletUrl()}${WALLET_PROVIDER_DISCONNECT_URL}?${callbackUrl}`;
+        return true;
     }
 
     /**
      * Returns currently connected address. Empty string if not connected
      */
     async getAddress(): Promise<string> {
-        if (!this.mainFrame) {
-            return '';
-        }
-
-        const { contentWindow } = this.mainFrame;
-        if (!contentWindow) {
-            return '';
-        }
-
-        return new Promise((resolve, reject) => {
-            console.log("postMessage", DAPP_MESSAGE_GET_ADDRESS);
-
-            contentWindow.postMessage({
-                type: DAPP_MESSAGE_GET_ADDRESS,
-            }, this.walletUrl);
-
-            const timeout = setTimeout(_ => reject('window not responding'), 5000);
-
-            const getAddress = (ev: IDappMessageEvent) => {
-
-                if (!this.isValidWalletSource(ev.origin)) {
-                    return;
-                }
-
-                const { data } = ev;
-                if (data.type !== DAPP_MESSAGE_GET_ADDRESS) {
-                    return;
-                }
-
-                clearTimeout(timeout);
-                window.removeEventListener('message', getAddress.bind(this));
-                return resolve(data.data);
-            };
-
-            window.addEventListener('message', getAddress);
-        });
+        throw new ErrNotImplemented();
     }
 
     /**
@@ -244,54 +101,35 @@ export class WalletProvider implements IDappProvider {
      * @param options
      */
     async sendTransaction(transaction: Transaction, options?: { callbackUrl?: string }): Promise<Transaction> {
-        if (!this.mainFrame) {
-            throw new Error("Wallet provider is not initialised, call init() first");
-        }
+        let plainTransaction = WalletProvider.prepareWalletTransaction(transaction);
+        let url = `${this.baseWalletUrl()}${WALLET_PROVIDER_SEND_TRANSACTION_URL}?${this.buildTransactionUrl(plainTransaction)}`;
 
-        const { contentWindow } = this.mainFrame;
-        if (!contentWindow) {
-            throw new Error("Wallet provider is not initialised, call init() first");
-        }
+        window.location.href = `${url}&callbackUrl=${options !== undefined && options.callbackUrl !== undefined ? options.callbackUrl : window.location.href}`;
+        return transaction;
+    }
 
-        return new Promise<Transaction>((resolve, reject) => {
-            let plainTransaction = WalletProvider.prepareWalletTransaction(transaction);
-            console.log("postMessage", DAPP_MESSAGE_SEND_TRANSACTION_URL, plainTransaction);
-
-            contentWindow.postMessage({
-                type: DAPP_MESSAGE_SEND_TRANSACTION_URL,
-                data: {
-                    transaction: plainTransaction
-                }
-            }, this.walletUrl);
-
-            const timeout = setTimeout(_ => reject('send transaction url not responding'), 5000);
-            const sendTransactionUrl = (ev: IDappMessageEvent) => {
-                console.log("event", "sendTransactionUrl", ev);
-
-                if (!this.isValidWalletSource(ev.origin)) {
-                    return;
+    /**
+     * Packs an array of {$link Transaction} and redirects to the correct transaction sigining hook
+     *
+     * @param transactions
+     * @param options
+     */
+    async signTransactions(transactions: Transaction[], options?: { callbackUrl?: string }): Promise<Transaction[]> {
+        const jsonToSend: any = {};
+        transactions.map(tx => {
+            let plainTx =  WalletProvider.prepareWalletTransaction(tx);
+            for ( let txProp in plainTx ) {
+                if (plainTx.hasOwnProperty(txProp) && !jsonToSend.hasOwnProperty(txProp)) {
+                    jsonToSend[txProp] = [];
                 }
 
-                const { data } = ev;
-                if (data.type !== DAPP_MESSAGE_SEND_TRANSACTION_URL) {
-                    return;
-                }
-
-                clearTimeout(timeout);
-                window.removeEventListener('message', sendTransactionUrl.bind(this));
-
-                if (data.error) {
-                    return reject(data.error);
-                }
-
-                return resolve(data.data.toString());
-            };
-
-            window.addEventListener('message', sendTransactionUrl);
-        }).then((url: any) => {
-            window.location.href = `${this.baseWalletUrl()}${url}&callbackUrl=${options !== undefined && options.callbackUrl !== undefined ? options.callbackUrl : window.location.href}`;
-            return transaction;
+                jsonToSend[txProp].push(plainTx[txProp]);
+            }
         });
+
+        let url = `${this.baseWalletUrl()}${WALLET_PROVIDER_SIGN_TRANSACTION_URL}?${qs.stringify(jsonToSend)}`;
+        window.location.href = `${url}&callbackUrl=${options !== undefined && options.callbackUrl !== undefined ? options.callbackUrl : window.location.href}`;
+        return transactions;
     }
 
     /**
@@ -301,54 +139,21 @@ export class WalletProvider implements IDappProvider {
      * @param options
      */
     async signTransaction(transaction: Transaction, options?: { callbackUrl?: string }): Promise<Transaction> {
-        if (!this.mainFrame) {
-            throw new Error("Wallet provider is not initialised, call init() first");
+        let plainTransaction = WalletProvider.prepareWalletTransaction(transaction);
+        let url = `${this.baseWalletUrl()}${WALLET_PROVIDER_SIGN_TRANSACTION_URL}?${this.buildTransactionUrl(plainTransaction)}`;
+
+        window.location.href = `${url}&callbackUrl=${options !== undefined && options.callbackUrl !== undefined ? options.callbackUrl : window.location.href}`;
+        return transaction;
+    }
+
+    getTransactionsFromWalletUrl(): Transaction[] {
+        const transactions: Transaction[] = [];
+        const urlParams = qs.parse(window.location.search.slice(1));
+        if (!WalletProvider.isTxSignReturnSuccess(urlParams)) {
+            return transactions;
         }
 
-        const { contentWindow } = this.mainFrame;
-        if (!contentWindow) {
-            throw new Error("Wallet provider is not initialised, call init() first");
-        }
-
-        return new Promise<Transaction>((resolve, reject) => {
-            let plainTransaction = WalletProvider.prepareWalletTransaction(transaction);
-            console.log("postMessage", DAPP_MESSAGE_SIGN_TRANSACTION_URL, plainTransaction);
-
-            contentWindow.postMessage({
-                type: DAPP_MESSAGE_SIGN_TRANSACTION_URL,
-                data: {
-                    transaction: plainTransaction
-                }
-            }, this.walletUrl);
-
-            const timeout = setTimeout(_ => reject('sign transaction url not responding'), 5000);
-            const signTransactionUrl = (ev: IDappMessageEvent) => {
-                console.log("event", "signTransactionUrl", ev);
-
-                if (!this.isValidWalletSource(ev.origin)) {
-                    return;
-                }
-
-                const { data } = ev;
-                if (data.type !== DAPP_MESSAGE_SIGN_TRANSACTION_URL) {
-                    return;
-                }
-
-                clearTimeout(timeout);
-                window.removeEventListener('message', signTransactionUrl.bind(this));
-
-                if (data.error) {
-                    return reject(data.error);
-                }
-
-                return resolve(data.data.toString());
-            };
-
-            window.addEventListener('message', signTransactionUrl);
-        }).then((url: any) => {
-            window.location.href = `${this.baseWalletUrl()}${url}&callbackUrl=${options !== undefined && options.callbackUrl !== undefined ? options.callbackUrl : window.location.href}`;
-            return transaction;
-        });
+        return WalletProvider.getTxSignReturnValue(urlParams);
     }
 
     /**
@@ -357,6 +162,47 @@ export class WalletProvider implements IDappProvider {
      */
     async signMessage(_: SignableMessage): Promise<SignableMessage> {
         throw new ErrNotImplemented();
+    }
+
+    static isTxSignReturnSuccess(urlParams: any): boolean {
+        return urlParams.hasOwnProperty(WALLET_PROVIDER_CALLBACK_PARAM) && urlParams[WALLET_PROVIDER_CALLBACK_PARAM] === WALLET_PROVIDER_CALLBACK_PARAM_TX_SIGNED;
+    }
+
+    static getTxSignReturnValue(urlParams: any): Transaction[] {
+        const expectedProps = ["nonce", "value", "receiver", "sender", "gasPrice",
+            "gasLimit", "data", "chainID", "version", "signature"];
+
+        for (let txProp of expectedProps) {
+            if (!urlParams[txProp] || !Array.isArray(urlParams[txProp])) {
+                throw new ErrInvalidTxSignReturnValue();
+            }
+        }
+
+        const expectedLength = urlParams["nonce"].length;
+        for (let txProp of expectedProps) {
+            if (urlParams[txProp].length !== expectedLength) {
+                throw new ErrInvalidTxSignReturnValue();
+            }
+        }
+
+        const transactions: Transaction[] = [];
+        for (let i = 0; i < expectedLength; i++) {
+            let tx = new Transaction({
+                nonce: new Nonce(urlParams["nonce"][i]),
+                value: Balance.fromString(<string>urlParams["value"][i]),
+                receiver: Address.fromString(<string>urlParams["receiver"][i]),
+                gasPrice: new GasPrice(parseInt(<string>urlParams["gasPrice"][i])),
+                gasLimit: new GasLimit(parseInt(<string>urlParams["gasLimit"][i])),
+                data: new TransactionPayload(<string>urlParams["data"][i]),
+                chainID: new ChainID(<string>urlParams["chainID"][i]),
+                version: new TransactionVersion(parseInt(<string>urlParams["version"][i])),
+
+            });
+            tx.applySignature(new Signature(<string>urlParams["signature"][i]), Address.fromString(<string>urlParams["sender"][i]));
+            transactions.push(tx);
+        }
+
+        return transactions;
     }
 
     static prepareWalletTransaction(transaction: Transaction): any {
@@ -372,56 +218,22 @@ export class WalletProvider implements IDappProvider {
         return plainTransaction;
     }
 
-    private async waitForRemote(): Promise<boolean> {
-        return new Promise((resolve, reject) => {
-            const timeout = setTimeout(_ => reject(false), DAPP_DEFAULT_TIMEOUT);
-            const setConnected = (ev: IDappMessageEvent) => {
-
-                if (!this.mainFrame) {
-                    return;
-                }
-                if (!this.isValidWalletSource(ev.origin)) {
-                    return;
-                }
-                const { data } = ev;
-                if (data.type !== DAPP_MESSAGE_INIT) {
-                    return;
-                }
-
-                clearTimeout(timeout);
-                window.removeEventListener('message', setConnected);
-                this.mainFrame.dataset.initialized = data.data.toString();
-
-                resolve(data.data);
-            };
-            window.addEventListener('message', setConnected.bind(this));
-        });
-    }
-
-    private attachMainFrame() {
-        let currentMainframe: HTMLIFrameElement|null;
-        if (this.customId) {
-            currentMainframe = document.querySelector(`iframe[id="${this.customId}"]`);
-        } else {
-            currentMainframe = document.querySelector(`iframe[src="${this.walletUrl}"]`);
+    private buildTransactionUrl(transaction: TransactionMessage): string {
+        let urlString = `receiver=${transaction.receiver}&value=${transaction.value}`;
+        if (transaction.gasLimit) {
+            urlString += `&gasLimit=${transaction.gasLimit}`;
         }
-        if (currentMainframe) {
-            this.mainFrame = currentMainframe;
-            return;
+        if (transaction.gasPrice) {
+            urlString += `&gasPrice=${transaction.gasPrice}`;
+        }
+        if (transaction.data) {
+            urlString += `&data=${transaction.data}`;
+        }
+        if (transaction.nonce) {
+            urlString += `&nonce=${transaction.nonce}`;
         }
 
-        let mainFrame = document.createElement('iframe');
-        mainFrame.src = this.walletUrl;
-        mainFrame.id = this.customId;
-        Object.assign(mainFrame.style, mainFrameStyle());
-        mainFrame.dataset.initialized = "false";
-        document.body.appendChild(mainFrame);
-
-        this.mainFrame = mainFrame;
-    }
-
-    private isValidWalletSource(origin: string): boolean {
-        return origin === this.walletUrl || origin === this.baseWalletUrl();
+        return urlString;
     }
 
     private baseWalletUrl(): string {